refactor: clean up code and remove chatty logs at startup
