+++ coderd/coderd.go
@@ -568,14 +568,13 @@ func New(options *Options) *API {
 	// bugs that may only occur when a key isn't precached in tests and the latency cost is minimal.
 	cryptokeys.StartRotator(ctx, options.Logger, options.Database)
 
-	// Ensure system role permissions are current for all organizations.
-	// This backfills permissions for roles created by migration and
-	// updates permissions when new RBAC resources are added, while
-	// using advisory lock for multi-replica safety.
-	err = ReconcileOrgMemberRoles(ctx, options.Logger, options.Database)
+	// Ensure org-member system role permissions are current for all
+	// organizations.
+	//nolint:gocritic // We need to manage system roles
+	err = rolestore.ReconcileOrgMemberRoles(dbauthz.AsSystemRestricted(ctx), options.Logger, options.Database)
 	if err != nil {
-		// TODO:(geokat) Not using fatal here and just continuing after
-		// logging the error would be a potential security hole, right?
+		// Not ideal, but not using Fatal here and just continuing
+		// after logging the error would be a potential security hole.
 		options.Logger.Fatal(ctx, "failed to reconcile orgMember role permissions", slog.Error(err))
 	}
