+++ enterprise/coderd/organizations.go
@@ -12,11 +12,9 @@ import (
 	"github.com/coder/coder/v2/coderd/audit"
 	"github.com/coder/coder/v2/coderd/database"
 	"github.com/coder/coder/v2/coderd/database/db2sdk"
-	"github.com/coder/coder/v2/coderd/database/dbauthz"
 	"github.com/coder/coder/v2/coderd/database/dbtime"
 	"github.com/coder/coder/v2/coderd/httpapi"
 	"github.com/coder/coder/v2/coderd/httpmw"
-	"github.com/coder/coder/v2/coderd/rbac"
 	"github.com/coder/coder/v2/coderd/rbac/rolestore"
 	"github.com/coder/coder/v2/codersdk"
 )
@@ -285,23 +283,12 @@ func (api *API) postOrganizations(rw http.ResponseWriter, r *http.Request) {
 			return xerrors.Errorf("create organization: %w", err)
 		}
 
-		// Create the organization-member system role for this organization.
+		// Create the org-member system role for this organization.
 		// New organizations have workspace sharing enabled by default.
-		orgPerms, memberPerms := rbac.OrgMemberPermissions(false)
-		//nolint:gocritic // We need to create a system role
-		sysCtx := dbauthz.AsSystemRestricted(ctx)
-		_, err = tx.InsertCustomRole(sysCtx, database.InsertCustomRoleParams{
-			Name:              rbac.RoleOrgMember(),
-			DisplayName:       "",
-			OrganizationID:    uuid.NullUUID{UUID: organization.ID, Valid: true},
-			SitePermissions:   database.CustomRolePermissions{},
-			OrgPermissions:    rolestore.ConvertPermissionsToDB(orgPerms),
-			UserPermissions:   database.CustomRolePermissions{},
-			MemberPermissions: rolestore.ConvertPermissionsToDB(memberPerms),
-			IsSystem:          true,
-		})
+		err = rolestore.CreateOrgMemberRole(ctx, tx, organization)
 		if err != nil {
-			return xerrors.Errorf("create organization member system role: %w", err)
+			return xerrors.Errorf("create org-member role for organization %s: %w",
+				organization.ID, err)
 		}
 
 		_, err = tx.InsertOrganizationMember(ctx, database.InsertOrganizationMemberParams{
